Chat/pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = ""
authors = ["blacksmokezip <hanzomainer1488@gmail.com>"]
readme = "README.md"
packages = [{include = "chat"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
requests = "^2.32.2"
python-dotenv = "^1.0.1"
flask-socketio = "^5.3.6"
flask-cors = "^4.0.1"
eventlet = "^0.36.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
